Functions

Functions

Below you will find all methods exported by MonteCarloObservable.jl.

Index

MonteCarloObservable.Observable

Base.Distributed.clear!

MonteCarloObservable.add!

MonteCarloObservable.add!

MonteCarloObservable.inmemory

MonteCarloObservable.name

MonteCarloObservable.rename

MonteCarloObservable.reset!

MonteCarloObservable.timeseries

Docs

Base.Distributed.clear! — Method.

clear!(obs::Observable{T})

Clears all measurement information in obs. Identical to init! and reset!.

source

MonteCarloObservable.add! — Method.

add!(obs::Observable{T}, measurements::AbstractArray{T}; verbose=false)

Add multiple measurements to observable obs.

source

MonteCarloObservable.add! — Method.

add!(obs::Observable{T}, measurement::T; verbose=false)

Add a measurement to observable obs.

source

MonteCarloObservable.inmemory — Method.

inmemory(obs::Observable{T})

Checks wether the observable is kept in memory (vs. on disk).

source

MonteCarloObservable.name — Method.

name(obs::Observable{T})

Returns the name of the observable.

source

MonteCarloObservable.rename — Method.

rename(obs::Observable{T}, name)

Renames the observable.

source

MonteCarloObservable.reset! — Method.

reset!(obs::Observable{T})

Resets all measurement information in obs. Identical to init! and clear!.

source

MonteCarloObservable.timeseries — Method.

timeseries(obs::Observable{T})

Returns the measurement timeseries of an observable.

If keep_in_memory == false it will read the timeseries from disk and thus might take some time.

See also getindex and view.

source

MonteCarloObservable.Observable — Method.

Observable{T}(name)

Create an observable.

source