General

Methods: General

Below you find all general exports.

Index

MonteCarloObservable.Observable

MonteCarloObservable.Observable

Base.Distributed.clear!

Base.eltype

Base.getindex

Base.isempty

Base.length

Base.ndims

Base.push!

Base.push!

Base.size

Base.view

MonteCarloObservable.add!

MonteCarloObservable.add!

MonteCarloObservable.inmemory

MonteCarloObservable.name

MonteCarloObservable.rename

MonteCarloObservable.reset!

MonteCarloObservable.timeseries

Documentation

MonteCarloObservable.Observable — Method.

Observable(T::DataType, name::String; keyargs...)

Create an observable of type T.

See Observable{T}.

source

MonteCarloObservable.Observable — Method.

Observable{T}(name::String; keyargs...)

Create an observable of type T.

The following keywords are allowed:

alloc: preallocated size of time series container

outfile: default HDF5/JLD output file for io operations

dataset: target path within outfile

inmemory: wether to keep the time series in memory or on disk

See also Observable.

source

Base.Distributed.clear! — Method.

clear!(obs::Observable{T})

Clears all measurement information in obs. Identical to init! and reset!.

source

Base.eltype — Method.

eltype(obs::Observable{T})

Returns the type T of a measurment of the observable.

source

Base.getindex — Method.

getindex(obs::Observable{T}, args...)

Get an element of the measurement time series of the observable.

source

Base.isempty — Method.

isempty(obs::Observable{T})

Determine wether the observable hasn't been measured yet.

source

Base.length — Method.

length(obs::Observable{T})

Number of measurements of the observable.

source

Base.ndims — Method.

ndims(obs::Observable{T})

Number of dimensions of the observable (of one measurement).

Equivalent to ndims(T).

source

Base.push! — Method.

push!(obs::Observable{T}, measurements::AbstractArray{T}; verbose=false)

Add multiple measurements to observable obs. Note that because of preallocation this isn't really a push.

source

Base.push! — Method.

push!(obs::Observable{T}, measurement::T; verbose=false)

Add a measurement to observable obs. Note that because of preallocation this isn't really a push.

source

Base.size — Method.

size(obs::Observable{T})

Size of the observable (of one measurement).

source

Base.view — Method.

view(obs::Observable{T}, args...)

Get a view into the measurement time series of the observable.

source

MonteCarloObservable.add! — Method.

add!(obs::Observable{T}, measurements::AbstractArray{T}; verbose=false)

Add multiple measurements to observable obs.

source

MonteCarloObservable.add! — Method.

add!(obs::Observable{T}, measurement::T; verbose=false)

Add a measurement to observable obs.

source

MonteCarloObservable.inmemory — Method.

inmemory(obs::Observable{T})

Checks wether the observable is kept in memory (vs. on disk).

source

MonteCarloObservable.name — Method.

name(obs::Observable{T})

Returns the name of the observable.

source

MonteCarloObservable.rename — Method.

rename(obs::Observable{T}, name)

Renames the observable.

source

MonteCarloObservable.reset! — Method.

reset!(obs::Observable{T})

Resets all measurement information in obs. Identical to init! and clear!.

source

MonteCarloObservable.timeseries — Method.

timeseries(obs::Observable{T})

Returns the measurement time series of an observable.

If inmemory(obs) == false it will read the time series from disk and thus might take some time.

See also getindex and view.

source