IO

Methods: IO

Below you find all IO related exports.

Index

MonteCarloObservable.export_result

MonteCarloObservable.loadobs

MonteCarloObservable.loadobs_frommemory

MonteCarloObservable.saveobs

MonteCarloObservable.timeseries_frommemory

MonteCarloObservable.timeseries_frommemory_flat

Documentation

MonteCarloObservable.export_result — Method.

export_results(obs::Observable{T}[, filename::AbstractString, entryname::AbstractString; timeseries::Bool=false])

Export result for given observable nicely to JLD.

Will export name, number of measurements, estimates for mean and one-sigma error (standard deviation). Optionally (timeseries==true) exports the full time series as well.

source

MonteCarloObservable.loadobs — Method.

loadobs(filename::AbstractString, entryname::AbstractString)

Load complete representation of an observable from JLD file.

See also saveobs.

source

MonteCarloObservable.loadobs_frommemory — Method.

loadobs_frommemory(filename::AbstractString, group::AbstractString)

Create an observable based on memory dump (inmemory==false).

source

MonteCarloObservable.saveobs — Method.

saveobs(obs::Observable{T}[, filename::AbstractString, entryname::AbstractString])

Saves complete representation of the observable to JLD file.

Default filename is "Observables.jld" and default entryname is name(obs).

See also loadobs.

source

MonteCarloObservable.timeseries_frommemory — Method.

timeseries_frommemory(filename::AbstractString, group::AbstractString)

Load time series from memory dump (inmemory==false) in HDF5/JLD file.

Will load and concatenate time series chunks. Output will be a vector of measurements.

source

MonteCarloObservable.timeseries_frommemory_flat — Method.

timeseries_frommemory_flat(filename::AbstractString, group::AbstractString)

Load time series from memory dump (inmemory==false) in HDF5/JLD file.

Will load and concatenate time series chunks. Output will be higher-dimensional array whose last dimension corresponds to Monte Carlo time.

source