Plotting

Methods: Statistics

Below you find all plotting related exports.

Index

MonteCarloObservable.binningplot

MonteCarloObservable.corrplot

MonteCarloObservable.hist

PyPlot.plot

Documentation

MonteCarloObservable.binningplot — Method.

binningplot(obs::Observable{T})

Creates a plot of the binning error coefficient R as a function of bin size.

The coefficient R should (up to statistical fluctuations) show a plateau for larger bin sizes, indicating that the bin averages have become independent. For correlated data one has R>≈1 and sqrt(R) quantifies how much one would have underestimated the one-sigma errorbar.

See binning_error.

source

MonteCarloObservable.corrplot — Method.

corrplot(obs::Observable{T})

Plot the autocorrelation function of the observable.

source

MonteCarloObservable.hist — Method.

hist(obs::Observable{T}[; errors=true, digits=3])

Plot a histogram of the observable's time series.

source

PyPlot.plot — Method.

plot(obs::Observable{T}[; errors=true, digits=3])

Plot the observable's time series.

source