Statistics

Methods: Statistics

Below you find all statistics related exports.

Index

Base.mean

Base.std

Base.var

ErrorAnalysis.binning_error

ErrorAnalysis.jackknife_error

Documentation

Base.mean — Method.

mean(obs::Observable{T})

Estimate of the mean of the observable.

source

Base.std — Method.

std(obs::Observable{T})

Estimate of the standard deviation (one-sigma error) of the mean. Respects correlations between measurements through binning analysis.

Note that this is not the same as Base.std(timeseries(obs)), not even for uncorrelated measurements.

Corresponds to the square root of var(obs). See also mean(obs).

source

Base.var — Method.

var(obs::Observable{T})

Estimate of the variance of the mean. Respects correlations between measurements through binning analysis.

Note that this is not the same as Base.var(timeseries(obs)), not even for uncorrelated measurements.

Corresponds to the square of std(obs). See also mean(obs).

source

ErrorAnalysis.binning_error — Method.

binning_error(obs::Observable{T}[; binsize=0, warnings=false])

Calculates statistical one-sigma error (eff. standard deviation) for correlated data. How: Binning of data and assuming statistical independence of bins (i.e. R plateau has been reached). (Eq. 3.18 of Book basically)

The default binsize=0 indicates automatic binning.

source

ErrorAnalysis.jackknife_error — Method.

jackknife_error(g::Function, obs::Observable{T}; [binsize=10])

Computes the jackknife standard deviation of g(<obs>) by binning the observable's time series and performing leave-one-out analysis.

source