Functions

Functions

Below you will find all methods exported by MonteCarloObservable.jl.

Index

MonteCarloObservable.Observable

Base.Distributed.clear!

MonteCarloObservable.add!

MonteCarloObservable.add!

MonteCarloObservable.reset!

MonteCarloObservable.timeseries

Docs

Base.Distributed.clear! — Method.

clear!(obs::Observable{T})

Clears all measurement information in obs. Identical to init! and reset!.

source

MonteCarloObservable.add! — Method.

add!(obs::Observable{T}, measurements::AbstractArray{T}; verbose=false)

Add multiple measurements to observable obs.

source

MonteCarloObservable.add! — Method.

add!(obs::Observable{T}, measurement::T; verbose=false)

Add a measurement to observable obs.

source

MonteCarloObservable.reset! — Method.

reset!(obs::Observable{T})

Resets all measurement information in obs. Identical to init! and clear!.

source

MonteCarloObservable.timeseries — Method.

timeseries(obs::Observable{T})

Returns the measurement timeseries of this observable. If keep_in_memory == false it will read the timeseries from disk and thus might take some time.

See also getindex and view.

source

MonteCarloObservable.Observable — Method.

Observable{T}(name)

Create an observable.

source