Functions

Functions

Below you will find all methods exported by MonteCarloObservable.jl.

Index

MonteCarloObservable.Observable

Base.Distributed.clear!

Base.eltype

Base.getindex

Base.isempty

Base.length

Base.mean

Base.ndims

Base.push!

Base.push!

Base.size

Base.std

Base.var

Base.view

MonteCarloObservable.add!

MonteCarloObservable.add!

MonteCarloObservable.inmemory

MonteCarloObservable.loadobs

MonteCarloObservable.name

MonteCarloObservable.rename

MonteCarloObservable.reset!

MonteCarloObservable.saveobs

MonteCarloObservable.timeseries

Docs

Base.Distributed.clear! — Method.

clear!(obs::Observable{T})

Clears all measurement information in obs. Identical to init! and reset!.

source

Base.eltype — Method.

eltype(obs::Observable{T})

Returns the type T of a measurment of the observable.

source

Base.getindex — Method.

getindex(obs::Observable{T}, args...)

Get an element of the measurement timeseries of the observable.

source

Base.isempty — Method.

isempty(obs::Observable{T})

Determine wether the observable hasn't been measured yet.

source

Base.length — Method.

length(obs::Observable{T})

Number of measurements of the observable.

source

Base.mean — Method.

mean(obs::Observable{T})

Estimate of the mean of the observable.

source

Base.ndims — Method.

ndims(obs::Observable{T})

Number of dimensions of the observable (of one measurement).

Equivalent to ndims(T).

source

Base.push! — Method.

push!(obs::Observable{T}, measurements::AbstractArray{T}; verbose=false)

Add multiple measurements to observable obs. Note that because of preallocation this isn't really a push.

source

Base.push! — Method.

push!(obs::Observable{T}, measurement::T; verbose=false)

Add a measurement to observable obs. Note that because of preallocation this isn't really a push.

source

Base.size — Method.

size(obs::Observable{T})

Size of the observable (of one measurement).

source

Base.std — Method.

std(obs::Observable{T})

Estimate of the standard deviation (one-sigma error) of the mean. Respects correlations between measurements through binning analysis.

Note that this is not the same as Base.std(timeseries(obs)), not even for uncorrelated measurements.

Corresponds to the square root of var(obs). See also mean(obs).

source

Base.var — Method.

var(obs::Observable{T})

Estimate of the variance of the mean. Respects correlations between measurements through binning analysis.

Note that this is not the same as Base.var(timeseries(obs)), not even for uncorrelated measurements.

Corresponds to the square of std(obs). See also mean(obs).

source

Base.view — Method.

view(obs::Observable{T}, args...)

Get a view into the measurement timeseries of the observable.

source

MonteCarloObservable.add! — Method.

add!(obs::Observable{T}, measurements::AbstractArray{T}; verbose=false)

Add multiple measurements to observable obs.

source

MonteCarloObservable.add! — Method.

add!(obs::Observable{T}, measurement::T; verbose=false)

Add a measurement to observable obs.

source

MonteCarloObservable.inmemory — Method.

inmemory(obs::Observable{T})

Checks wether the observable is kept in memory (vs. on disk).

source

MonteCarloObservable.loadobs — Method.

loadobs(filename::AbstractString, entryname::AbstractString)

Load complete representation of an observable from JLD file.

See also saveobs.

source

MonteCarloObservable.name — Method.

name(obs::Observable{T})

Returns the name of the observable.

source

MonteCarloObservable.rename — Method.

rename(obs::Observable{T}, name)

Renames the observable.

source

MonteCarloObservable.reset! — Method.

reset!(obs::Observable{T})

Resets all measurement information in obs. Identical to init! and clear!.

source

MonteCarloObservable.saveobs — Method.

saveobs(obs::Observable{T}[, filename::AbstractString, entryname::AbstractString])

Saves complete representation of the observable to JLD file.

Default filename is "Observables.jld" and default entryname is name(obs).

See also loadobs.

source

MonteCarloObservable.timeseries — Method.

timeseries(obs::Observable{T})

Returns the measurement timeseries of an observable.

If inmemory(obs) == false it will read the timeseries from disk and thus might take some time.

See also getindex and view.

source

MonteCarloObservable.Observable — Method.

Observable{T}(name)

Create an observable.

source